[FIX/#11] activity -> fragment 로 수정
